refactor: styles in sidebar component
