+++ @ui/ui-kit/Icon/MakeIcon.tsx
@@ -4,10 +4,10 @@ import style from './IconsComponent.module.scss'
 import Link from 'next/link'
 
 export const MakeIcon: FC<MakeIconPropsType> = (props) => {
-  const { OutlineIcon, Icon, className, text, isActive, isDisabled, Url } = props
+  const { OutlineIcon, Icon, className, text, isActive, isDisabled, url } = props
   return (
     <Link
-      href={(Url && !isDisabled) ? Url : ''}
+      href={(url && !isDisabled) ? url : ''}
       className={`${style.IconWrapper} ${className} ${isActive && style.IconWrapperActive}
         ${isDisabled && style.IconWrapperDisabled}`}
     >
