--- public/icons/icons.ts
@@ -1,35 +1,0 @@
-import HomeOutline from '@/public/sidebar-icons/home-outline.svg'
-import Home from '@/public/sidebar-icons/home.svg'
-import CreateOutline from '@/public/sidebar-icons/plus-square-outline.svg'
-import Create from '@/public/sidebar-icons/plus-square.svg'
-import ProfileOutline from '@/public/sidebar-icons/person-outline.svg'
-import Profile from '@/public/sidebar-icons/person.svg'
-import MessageOutline from '@/public/sidebar-icons/message-circle-outline.svg'
-import Message from '@/public/sidebar-icons/message-circle.svg'
-import SearchOutline from '@/public/sidebar-icons/search-outline.svg'
-import Search from '@/public/sidebar-icons/search.svg'
-import TrendingUpOutline from '@/public/sidebar-icons/trending-up-outline.svg'
-import TrendingUp from '@/public/sidebar-icons/trending-up.svg'
-import BookmarkOutline from '@/public/sidebar-icons/bookmark-outline.svg'
-import Bookmark from '@/public/sidebar-icons/bookmark.svg'
-import LogoutOutline from '@/public/sidebar-icons/log-out-outline.svg'
-import Logout from '@/public/sidebar-icons/log-out.svg'
-
-export const icons = {
-HomeOutline,
-Home,
-CreateOutline,
-Create,
-ProfileOutline,
-Profile,
-MessageOutline,
-Message,
-SearchOutline,
-Search,
-TrendingUpOutline,
-TrendingUp,
-BookmarkOutline,
-Bookmark,
-LogoutOutline,
-Logout,
-}
