+++ public/sidebar-icons/icons.ts
@@ -1,0 +1,35 @@
+import HomeOutline from './home-outline.svg'
+import Home from './home.svg'
+import CreateOutline from './plus-square-outline.svg'
+import Create from './plus-square.svg'
+import ProfileOutline from './person-outline.svg'
+import Profile from './person.svg'
+import MessageOutline from './message-circle-outline.svg'
+import Message from './message-circle.svg'
+import SearchOutline from './search-outline.svg'
+import Search from './search.svg'
+import TrendingUpOutline from './trending-up-outline.svg'
+import TrendingUp from './trending-up.svg'
+import BookmarkOutline from './bookmark-outline.svg'
+import Bookmark from './bookmark.svg'
+import LogoutOutline from './log-out-outline.svg'
+import Logout from './log-out.svg'
+
+export const icons = {
+HomeOutline,
+Home,
+CreateOutline,
+Create,
+ProfileOutline,
+Profile,
+MessageOutline,
+Message,
+SearchOutline,
+Search,
+TrendingUpOutline,
+TrendingUp,
+BookmarkOutline,
+Bookmark,
+LogoutOutline,
+Logout,
+}
